Update opening in external tab for key auth
I was previously using cookies with SameSite=Lax. However, this
doesn't seem to be working anymore, possibly because of an update
to Chrome or perhaps I'm doing something differently now.

Either way, I've decided to avoid cookies entirely for the time
being, since there are lots of security issues anyway.

So this commit simply generates a one-off key and adds it to the
URL.